+++ script/tool/CHANGELOG.md
@@ -9,6 +9,7 @@
   immediately abort the test.
 - Deprecated `--plugins` in favor of new `--packages`. `--plugins` continues to
   work for now, but will be removed in the future.
+- Make `drive-examples` device detection robust against Flutter tool banners.
 
 ## 0.3.0